tests/wpt/web-platform-tests/html/browsers/browsing-the-web/history-traversal/unset_context_name-1.html

<!doctype html>
<title>window.name after navigating to a different origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {assert_equals(window.name, "")})
</script>
